Pull-up operation async check to upper class
